fix: different render approach
